Split endpoint error helpers
